Illustrate sparse topology
The diagram captures a transitional state where SPT has been formed
with assistance of RP, followed by the severing of the link between
rtr1<->rtr2. After this the SPT becomes the RPT because traffic can
only flow via the RP node. After convergence (removal of LLDP
neighborship) the rtr1<->rtr2 link was removed altogether from the
diagram, so I wanted to include this preliminary state instead
because it shows the full topology.

IIUC the grapher currently does not draw all LLDP-discovered links;
rather only the ones through which a join has occurred. Need to
modify the python code to amend the full topology.

Also the PIM join alongside the IGMP join from the Ubuntu host is a
bit of a synthetic hack at the moment.